Refactor move package
